Worked on prompts for viewing all departments, viewing all roles, viewing all employees, adding a department, and adding a role
