--- db/query.sql
@@ -1,4 +1,0 @@
-SELECT *
-FROM departments, roles, employees
-JOIN departments ON roles.departments_id = departments.id;
-JOIN roles ON employees.role_id = roles.id;
